member initialization
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@11933 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/components/ASSDriver/ASSCore/ASSDriver.h
@@ -147,7 +147,7 @@ class ASSDriver :
         RasterHandle allocateRaster(const std::string & theName);
         void readSensorValues();
         void processSensorValues( double theDeltaT);
-        uint16_t readStatusToken( std::vector<unsigned char>::iterator & theIt, const char theToken );
+        unsigned readStatusToken( std::vector<unsigned char>::iterator & theIt, const char theToken );
         void parseStatusLine();
         void updateDerivedRasters();
         void updateCursors( double theDeltaT);
@@ -176,7 +176,7 @@ class ASSDriver :
         y60::ScenePtr _myScene;
 
         asl::Time   _myLastFrameTime;
-        float       _myRunTime;
+        double       _myRunTime;
         int         _myComponentThreshold;
         int         _myNoiseThreshold;
         float       _myIntensityThreshold;
